refactor navigation into separate nav & aside comps
