+++ src/App.vue
@@ -1,12 +1,7 @@
 <script setup lang="ts">
 import { RouterView } from 'vue-router'
-import Navigation from '@/components/Navigation.vue'
 </script>
 
 <template>
-  <Navigation>
-    <template #tab>
-      <RouterView />
-    </template>
-  </Navigation>
+  <RouterView />
 </template>
