+++ src/components/Navigation/NavBar.vue
@@ -1,0 +1,5 @@
+<template>
+  <nav class="px-4 flex items-center justify-between py-2 bg-dark-100 h-[var(--nav-height)]">
+    <div>TodoApp</div>
+  </nav>
+</template>
