+++ src/views/HomeView.vue
@@ -1,9 +1,14 @@
 <script setup lang="ts">
 import TasksList from '@/components/TasksList.vue'
+import Navigation from '@/components/Navigation/Index.vue'
 </script>
 
 <template>
-  <main class="h-full">
-    <TasksList />
-  </main>
+  <Navigation>
+    <template #tab>
+      <main class="h-full">
+        <TasksList />
+      </main>
+    </template>
+  </Navigation>
 </template>
